@@ -31,7 +31,8 @@
     "@mui/x-date-pickers": "^6.0.4",
     "@rails/actioncable": "^7.0.8",
     "@react-google-maps/api": "^2.12.2",
-    "@sentry/nextjs": "^7.60.0",
+    "@sentry/nextjs": "^7.91.0",
+    "@sentry/utils": "^7.91.0",
     "@storybook/react": "^6.5.16",
     "@tinymce/tinymce-react": "^4.3.0",
     "@types/actioncable": "^5.2.8",
@@ -52,19 +53,19 @@
     "lodash": "^4.17.21",
     "moment": "^2.29.4",
     "moment-timezone": "^0.5.42",
-    "next": "12.2.5",
+    "next": "^14.0.4",
     "next-transpile-modules": "^9.0.0",
     "qs": "^6.11.2",
     "query-string": "^7.1.1",
     "ra-core": "^4.3.1",
     "ra-data-json-server": "^4.3.1",
     "rc-slider": "^10.0.1",
-    "react": "18.2.0",
+    "react": "^18.2.0",
     "react-admin": "^4.3.1",
     "react-cookie": "^4.1.1",
     "react-custom-scrollbars-2": "^4.5.0",
     "react-dates": "^21.8.0",
-    "react-dom": "18.2.0",
+    "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.3",
     "react-ga4": "^2.1.0",
     "react-google-autocomplete": "^2.7.3",
@@ -72,10 +73,10 @@
     "react-i18next": "^13.2.2",
     "react-phone-input-2": "^2.15.1",
     "react-qr-code": "^2.0.12",
-    "react-relay": "^14.1.0",
+    "react-relay": "^16.1.0",
     "react-router-dom": "^6.3.0",
     "relay-nextjs": "^0.9.0",
-    "relay-runtime": "^15.0.0",
+    "relay-runtime": "^16.1.0",
     "request": "^2.88.2",
     "sonner": "^0.6.2",
     "styled-components": "^5.3.5",
@@ -101,8 +102,8 @@
     "@types/react": "18.0.18",
     "@types/react-dates": "^21.8.3",
     "@types/react-dom": "18.0.6",
-    "@types/react-relay": "^14.1.1",
-    "@types/relay-runtime": "^14.1.2",
+    "@types/react-relay": "^14.1.7",
+    "@types/relay-runtime": "^14.1.21",
     "@types/uuid": "^9.0.0",
     "@typescript-eslint/eslint-plugin": "^5.36.1",
     "@typescript-eslint/parser": "^5.36.1",
@@ -123,7 +124,7 @@
     "eslint-plugin-storybook": "^0.6.6",
     "graphql": "^16.6.0",
     "prettier": "^2.7.1",
-    "relay-compiler": "^14.1.0",
+    "relay-compiler": "^16.1.0",
     "relay-compiler-language-typescript": "^15.0.1",
     "relay-config": "^12.0.1"
   }
