@@ -1,43 +1,73 @@
 import React from "react";
-import { useRouter } from "next/router";
-import { QueryContext } from "../../components/QueryProvider";
+import { usePathname, useRouter, useSearchParams } from "next/navigation";
 
-export function useQuery(key?: string, defaultValue?: any) {
-  const router = useRouter();
-  const { query, setKey } = React.useContext(QueryContext);
-  React.useEffect(() => {
-    if (defaultValue && key) {
-      const isArray = Array.isArray(defaultValue);
-      const routerValue =
-        !Array.isArray(router.query[key]) && isArray
-          ? [router.query[key]].filter(Boolean)
-          : router.query[key];
-
-      setKey(key, routerValue || defaultValue);
-    }
-  }, []);
-
-  return React.useMemo(() => {
-    if (key) {
-      if (Object.keys(query).includes(key)) {
-        return query[key];
-      }
-      return defaultValue;
-    }
-    return query;
-  }, [query, key]);
+export function parseValue(value: string) {
+  try {
+    return JSON.parse(value);
+  } catch (error) {
+    console.warn(error);
+    return undefined;
+  }
 }
 
-export function useUpdateQuery(key: string) {
-  const { setKey, deleteKey, query } = React.useContext(QueryContext);
+export function useQuery(
+  key: string,
+  defaultValue?: any,
+  parse?: (value: string) => any
+) {
+  const query = useSearchParams();
+  const value = query.get(key);
+  return React.useMemo(() => {
+    const useCustomParse = parse instanceof Function;
+
+    if (key) {
+      if (!value) {
+        return typeof defaultValue === "undefined" ? null : defaultValue;
+      }
+
+      if (useCustomParse && value) {
+        return parse(value);
+      }
+
+      if (value) {
+        return parseValue(value);
+      }
+    }
+
+    return null;
+  }, [query, value, key]);
+}
+
+export function useUpdateQuery(key: string, format?: (value: any) => string) {
+  const router = useRouter();
+  const pathname = usePathname();
+  const searchParams = useSearchParams();
+  const query = new URLSearchParams(Array.from(searchParams.entries())); // -> has to use this form
+
   return React.useCallback(
     (value: any) => {
-      if (value) {
-        setKey(key, value);
+      const useCustomFormat = format instanceof Function;
+
+      if (!value) {
+        query.delete(key);
       } else {
-        deleteKey(key);
+        if (useCustomFormat) {
+          value = format(value);
+        } else {
+          value = JSON.stringify(value);
+        }
+
+        query.set(key, value);
       }
+
+      let queryString = query.toString();
+
+      if (!queryString) {
+        queryString = "";
+      }
+
+      router.push(`${pathname}?${queryString}`);
     },
-    [key, setKey, query]
+    [key, query, pathname, searchParams, router]
   );
 }
