@@ -2,7 +2,7 @@ import { useMemo } from "react";
 import { graphql } from "react-relay";
 import * as Yup from "yup";
 import { yupResolver } from "@hookform/resolvers/yup";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import { UseFormReturn } from "react-hook-form/dist/types";
 import { UseMutationConfig } from "react-relay/relay-hooks/useMutation";
 import { Disposable } from "relay-runtime";
@@ -72,7 +72,7 @@ export function useSignInForm(onNextStep: (delay: number) => void) {
 
             toast.success(result.signIn?.notification!);
           } else if (result.signIn?.user?.id) {
-            router.push("/events");
+            router.push("/trips");
 
             toast.success(result.signIn?.notification!);
           }
