@@ -1,10 +1,13 @@
 import React from "react";
 import moment from "moment";
 import { useTranslation } from "react-i18next";
-import { getHumanDateTime } from "../../../../lib/utils/dates";
-import useStatusColor from "../../../../lib/hooks/useStatusColor";
-import Tag from "../../../v2/Tag/Tag";
-import Link from "../../../v2/Link";
+import { graphql, useFragment } from "react-relay";
+import { getHumanDateTime } from "lib/utils/dates";
+import useStatusColor from "lib/hooks/useStatusColor";
+import Tag from "components/v2/Tag/Tag";
+import Link from "components/v2/Link";
+import { schedules_useSchedulesColumns_SchedulesConnectionFragment$key } from "artifacts/schedules_useSchedulesColumns_SchedulesConnectionFragment.graphql";
+import useEdges from "../../../../lib/hooks/useEdges";
 
 const TagColor = ({ status }: { status: string }) => {
   const color = useStatusColor({
@@ -24,6 +27,11 @@ export function useSchedulesHeaders() {
   const { t } = useTranslation();
   return React.useMemo(
     () => [
+      {
+        label: t("models.schedule.attributes.id"),
+        width: 150,
+        key: "id",
+      },
       {
         label: t("models.event.singular"),
         width: 300,
@@ -55,31 +63,62 @@ export function useSchedulesHeaders() {
 }
 
 export function useSchedulesColumns(
-  schedules: ReadonlyArray<Record<string, any>>
+  schedulesConnectionRef: schedules_useSchedulesColumns_SchedulesConnectionFragment$key
 ) {
+  const schedules =
+    useFragment<schedules_useSchedulesColumns_SchedulesConnectionFragment$key>(
+      graphql`
+        fragment schedules_useSchedulesColumns_SchedulesConnectionFragment on ScheduleConnection {
+          edges {
+            node {
+              id
+              scheduledFor
+              status
+              bookedPlaces
+              bookings {
+                nodes {
+                  id
+                }
+              }
+              event {
+                id
+                title
+              }
+            }
+          }
+        }
+      `,
+      schedulesConnectionRef
+    );
+  const scheduleNodes = useEdges(schedules);
   return React.useMemo(
     () =>
-      schedules.map((scheduleRow) => ({
-        date: getHumanDateTime(moment(scheduleRow.scheduledFor)),
-        bookings: scheduleRow.bookings?.length,
-        attendees: scheduleRow.bookings?.reduce(
-          (acc: number, booking: Record<string, any>) => {
-            acc += booking.attendees.length;
-            return acc;
-          },
-          0
-        ),
-        status: <TagColor status={scheduleRow.status} />,
-        eventId: (
-          <Link
-            href={`/my-firm/events/${scheduleRow.event?.id}`}
-            primary
-            fontSize="sm"
-          >
-            {scheduleRow.event?.title}
-          </Link>
-        ),
-      })),
-    [schedules]
+      scheduleNodes
+        .map((v) => v!)
+        .map((scheduleRow) => ({
+          id: (
+            <Link
+              href={`/my-firm/schedules/${scheduleRow.id}`}
+              primary
+              fontSize="sm"
+            >
+              {scheduleRow.id}
+            </Link>
+          ),
+          date: getHumanDateTime(moment(scheduleRow.scheduledFor)),
+          bookings: scheduleRow.bookings.nodes.length,
+          attendees: scheduleRow.bookedPlaces,
+          status: <TagColor status={scheduleRow.status} />,
+          eventId: (
+            <Link
+              href={`/my-firm/events/${scheduleRow.event?.id}`}
+              primary
+              fontSize="sm"
+            >
+              {scheduleRow.event?.title}
+            </Link>
+          ),
+        })),
+    [scheduleNodes]
   );
 }
