@@ -1,10 +1,25 @@
 // Injected content via Sentry wizard below
 const { withSentryConfig } = require("@sentry/nextjs");
+const path = require("path");
 
 /** @type {import('next').NextConfig} */
 
 const nextConfig = {
     reactStrictMode: true,
+    experimental: {
+        windowHistorySupport: true
+    },
+    compiler: {
+        relay: {
+            // This should match relay.config.js
+            artifactDirectory: './artifacts',
+            schema: path.resolve(__dirname, './schema.graphql'),
+            src: path.resolve(__dirname),
+            exclude: ['**/node_modules/**', '**/__generated__/**', 'schema.graphql'],
+            language: "typescript",
+            noFutureProofEnums: true,
+        },
+    },
     images: {
         domains: ['placehold.co', 's3.eu-north-1.amazonaws.com'],
     },
@@ -12,7 +27,6 @@ const nextConfig = {
         GRAPHQL_API_URL: process.env.GRAPHQL_API_URL,
         SENTRY_DSN: process.env.SENTRY_DSN,
     },
-    webpack5: true,
     webpack: (config) => {
         config.resolve.fallback = { fs: false };
 
