@@ -1,30 +1,33 @@
 import React from "react";
-import { Autocomplete, Box, ButtonGroup, FormControl, FormLabel, Grid, IconButton, Option, Stack, Tooltip, useTheme } from "@mui/joy";
+import {
+  Autocomplete,
+  ButtonGroup,
+  FormControl,
+  FormLabel,
+  Grid,
+  IconButton,
+  Stack,
+  Tooltip,
+  useTheme,
+} from "@mui/joy";
 import { graphql, useFragment } from "react-relay";
 import moment, { Moment } from "moment";
 import { useTranslation } from "react-i18next";
-import DateCalendar from "../../../../../components/v2/DateCalendar/DateCalendar";
-import {
-  dateFormat,
-  setTime,
-  timeFormat,
-} from "../../../../../lib/utils/dates";
-import Input from "../../../../../components/v2/Input";
-import Select from "../../../../../components/v2/Select";
-import { getCurrencyFormat } from "../../../../../lib/utils/currencyFormatter";
-import Typography from "../../../../../components/v2/Typography";
-import Button from "../../../../../components/v2/Button";
-import useFormContext from "../../../../../lib/hooks/useFormContext";
-import useUniqueMomentDates from "../../../../../lib/hooks/useUniqueMomentDates";
-import useTimeFromDate from "../../../../../lib/hooks/useTimeFromDate";
-import Link from "../../../../../components/v2/Link";
-import { BookEvent_EventFragment$key } from "../../../../../artifacts/BookEvent_EventFragment.graphql";
-import SubmitButton from "../../../../../components/shared/SubmitButton";
-import { capitalize } from "../../../../../lib/utils/capitalize";
 import AddIcon from "@mui/icons-material/Add";
 import RemoveIcon from "@mui/icons-material/Remove";
 import { useMediaQuery } from "@mui/material";
-import { isMobile } from "filestack-js";
+import DateCalendar from "components/v2/DateCalendar/DateCalendar";
+import { setTime, timeFormat } from "lib/utils/dates";
+import { getCurrencyFormat } from "lib/utils/currencyFormatter";
+import Typography from "components/v2/Typography";
+import Button from "components/v2/Button";
+import useFormContext from "lib/hooks/useFormContext";
+import useUniqueMomentDates from "lib/hooks/useUniqueMomentDates";
+import useTimeFromDate from "lib/hooks/useTimeFromDate";
+import Link from "components/v2/Link";
+import { BookEvent_EventFragment$key } from "artifacts/BookEvent_EventFragment.graphql";
+import SubmitButton from "components/shared/SubmitButton";
+import { capitalize } from "lib/utils/capitalize";
 
 interface BookEventProps {
   eventFragmentRef: BookEvent_EventFragment$key;
@@ -89,13 +92,20 @@ const BookEvent = ({ eventFragmentRef }: BookEventProps) => {
     moment(sch.scheduledFor).isSame(moment(dateField.value))
   );
   const { t } = useTranslation();
-  const theme = useTheme()
-  const isMobileView = useMediaQuery(theme.breakpoints.down('md'))
+  const theme = useTheme();
+  const isMobileView = useMediaQuery(theme.breakpoints.down("md"));
+
+  React.useEffect(() => {
+    const time = availableTimes[0]?.format(timeFormat);
+    if (availableTimes.length === 1 && time && selectedTime !== time) {
+      dateField.onChange(setTime(dateField.value, time));
+    }
+  }, [availableTimes, dateField]);
 
   return (
-    <Grid container justifyContent={'center'}>
+    <Grid container justifyContent="center">
       <Grid md={12} sm={12}>
-        <Stack spacing={1} useFlexGap alignItems={'center'}>
+        <Stack spacing={1} useFlexGap alignItems="center">
           <DateCalendar
             availableDates={availableDates}
             highlightedDates={bookedDates}
@@ -103,55 +113,70 @@ const BookEvent = ({ eventFragmentRef }: BookEventProps) => {
             disablePast
             sx={{
               maxWidth: "100%",
-              margin: isMobileView ? '0 autp' : 'unset',
-              padding: 0
+              margin: isMobileView ? "0 auto" : "unset",
+              padding: 0,
             }}
             onChange={(date) => {
               if (!date) return;
               dateField.onChange(date.startOf("day"));
             }}
           />
-          <Stack direction={'row'} spacing={1} useFlexGap justifyContent={'flex-end'}>
-            <FormControl sx={{ margin: 0, width: '100%' }} >
-              <FormLabel>{t('datepicker.selectTime')}</FormLabel>
+          <Stack
+            direction="row"
+            spacing={1}
+            useFlexGap
+            justifyContent="flex-end"
+          >
+            <FormControl sx={{ margin: 0, width: "100%" }}>
+              <FormLabel>{t("datepicker.selectTime")}</FormLabel>
               <Autocomplete
                 disableClearable
-                value={{ value: selectedTime, label: selectedTime }}
+                value={
+                  selectedTime
+                    ? { value: selectedTime, label: selectedTime }
+                    : { value: null, label: t("datepicker.selectTime") }
+                }
                 options={availableTimes.map((time: Moment) => ({
                   label: time.format(timeFormat),
-                  value: time.format(timeFormat)
+                  value: time.format(timeFormat),
                 }))}
-                onChange={(event, { value }) => {
+                onChange={(_, { value }) => {
                   if (!value) return;
 
                   dateField.onChange(setTime(dateField.value, value));
                 }}
-                sx={{ margin: 0, maxWidth: '125px' }}
-                size={'sm'}
+                sx={{ margin: 0, maxWidth: "125px" }}
+                size="sm"
               />
             </FormControl>
-            <FormControl sx={{margin: 0}}>
-              <FormLabel>{t('models.attendee.plural')}</FormLabel>
+            <FormControl sx={{ margin: 0 }}>
+              <FormLabel>{t("models.attendee.plural")}</FormLabel>
               <ButtonGroup>
-                <Tooltip title={t('forms.removeAttendee.action')}>
+                <Tooltip title={t("forms.removeAttendee.action")}>
                   <IconButton
-                    disabled={attendeesCountField.value.length === 1 || !!booking}
-                    onClick={() => attendeesCountField.onChange(attendeesCountField.value - 1)}
-                    size={'sm'}
+                    disabled={
+                      attendeesCountField.value.length === 1 || !!booking
+                    }
+                    onClick={() =>
+                      attendeesCountField.onChange(
+                        attendeesCountField.value - 1
+                      )
+                    }
+                    size="sm"
                   >
                     <RemoveIcon />
                   </IconButton>
                 </Tooltip>
-                <IconButton
-                  size={'sm'}
-                >
-                  {attendeesCountField.value}
-                </IconButton>
-                <Tooltip title={t('forms.addAttendee.action')}>
+                <IconButton size="sm">{attendeesCountField.value}</IconButton>
+                <Tooltip title={t("forms.addAttendee.action")}>
                   <IconButton
                     disabled={!!booking}
-                    onClick={() => attendeesCountField.onChange(attendeesCountField.value + 1)}
-                    size={'sm'}
+                    onClick={() =>
+                      attendeesCountField.onChange(
+                        attendeesCountField.value + 1
+                      )
+                    }
+                    size="sm"
                   >
                     <AddIcon />
                   </IconButton>
@@ -164,11 +189,13 @@ const BookEvent = ({ eventFragmentRef }: BookEventProps) => {
               event.attendeePricePerUom?.cents,
               event.attendeePricePerUom?.currency?.name
             )}{" "}
-            x {booking ? booking.attendees.length : attendeesCountField.value} {t("general.attendee")}
+            x {booking ? booking.attendees.length : attendeesCountField.value}{" "}
+            {t("general.attendee")}
             <br />
             {capitalize(t("general.total"))}:{" "}
             {getCurrencyFormat(
-              (booking ? booking.attendees.length : attendeesCountField.value) * (event.attendeePricePerUom?.cents || 0),
+              (booking ? booking.attendees.length : attendeesCountField.value) *
+                (event.attendeePricePerUom?.cents || 0),
               event.attendeePricePerUom?.currency?.name
             )}
           </Typography>
@@ -181,8 +208,13 @@ const BookEvent = ({ eventFragmentRef }: BookEventProps) => {
             </SubmitButton>
           )}
           {booking && (
-            <Link href={`/trips/${booking.trip.id}#${booking.id}`} underline={false}>
-              <Button fullWidth>{t('scenes.attendees.events.eventScene.details')}</Button>
+            <Link
+              href={`/trips/${booking.trip.id}#${booking.id}`}
+              underline={false}
+            >
+              <Button fullWidth>
+                {t("scenes.attendees.events.eventScene.details")}
+              </Button>
             </Link>
           )}
           {schedule?.leftPlaces && !booking && (
