@@ -7,17 +7,18 @@ import {
   useFragment,
   usePaginationFragment,
 } from "react-relay";
-import Sidebar from "./components/Sidebar";
-import SearchBar from "./components/SearchBar";
-import { EventsScene_EventsPaginationFragment$key } from "../../../../artifacts/EventsScene_EventsPaginationFragment.graphql";
-import EventCardCompact from "./components/EventCardCompact";
-import EventCardWide from "./components/EventCardWide";
+import { EventsScene_EventsPaginationFragment$key } from "artifacts/EventsScene_EventsPaginationFragment.graphql";
+import { usePagedEdges } from "lib/hooks/usePagedEdges";
+import { EventsScenePaginationQuery } from "artifacts/EventsScenePaginationQuery.graphql";
+import { EventsScene_InterestsFragment$key } from "artifacts/EventsScene_InterestsFragment.graphql";
+import { GlobalSidebarContext } from "components/GlobalSidebarProvider";
+import { parseValue, useQuery, useUpdateQuery } from "lib/hooks/useQuery";
+import moment, { Moment } from "moment";
 import Pagination from "./components/Pagination";
-import { usePagedEdges } from "../../../../lib/hooks/usePagedEdges";
-import { EventsScenePaginationQuery } from "../../../../artifacts/EventsScenePaginationQuery.graphql";
-import { EventsScene_InterestsFragment$key } from "../../../../artifacts/EventsScene_InterestsFragment.graphql";
-import { GlobalSidebarContext } from "../../../../components/GlobalSidebarProvider";
-import { useQuery, useUpdateQuery } from "../../../../lib/hooks/useQuery";
+import EventCardWide from "./components/EventCardWide";
+import EventCardCompact from "./components/EventCardCompact";
+import SearchBar from "./components/SearchBar";
+import Sidebar from "./components/Sidebar";
 
 interface Props {
   eventsFragmentRef:
@@ -37,8 +38,7 @@ const EventsScene = ({ eventsFragmentRef }: Props) => {
   const showSidebar = useMediaQuery(theme.breakpoints.up("md"));
   const isLargeDisplay = useMediaQuery(theme.breakpoints.up("lg"));
   const isVeryLargeDisplay = useMediaQuery(theme.breakpoints.up("xl"));
-  const currentPage = useQuery("currentPage", 1);
-  const updateCurrentPage = useUpdateQuery("currentPage");
+  const [currentPage, updateCurrentPage] = React.useState<number>(1);
   const { data, hasPrevious, hasNext, loadPrevious, loadNext, refetch } =
     usePaginationFragment<
       EventsScenePaginationQuery,
@@ -63,6 +63,14 @@ const EventsScene = ({ eventsFragmentRef }: Props) => {
           }
           eventFilters {
             ...Sidebar_EventFiltersFragment
+            startDate
+            endDate
+            minPrice {
+              cents
+            }
+            maxPrice {
+              cents
+            }
           }
         }
       `,
@@ -77,24 +85,34 @@ const EventsScene = ({ eventsFragmentRef }: Props) => {
     `,
     eventsFragmentRef as EventsScene_InterestsFragment$key
   );
-  const query = useQuery("query");
-  const interestsSlug = useQuery("interests");
+  const query = useQuery("query", "");
+  const maxPrice = useQuery(
+    "maxPrice",
+    data.eventFilters.maxPrice.cents / 100,
+    (value) => parseInt(value, 10)
+  );
+
+  const minPrice = useQuery(
+    "minPrice",
+    data.eventFilters.minPrice.cents / 100,
+    (value) => parseInt(value, 10)
+  );
+
+  const interests = useQuery("interests", [], (value) =>
+    Array.from(parseValue(value))
+  );
   const updateInterests = useUpdateQuery("interests");
+  const dates = useQuery("dates", [], (dts) =>
+    parseValue(dts)
+      .map((dt: string) => moment(dt))
+      .filter((dt: Moment) => dt.isValid)
+  );
+  const startDate = dates[0]?.toISOString();
+  const endDate = dates[1]?.toISOString();
   const events = usePagedEdges(data.events, currentPage, 10);
-  const [{ filters }, setFilters] = React.useState<any>({
-    filters: { query },
-  });
   const queryRef = React.useRef<Disposable>();
 
   React.useEffect(() => {
-    const startDate = filters?.startDate
-      ? filters?.startDate.toISOString()
-      : undefined;
-
-    const endDate = filters?.endDate
-      ? filters?.endDate.toISOString()
-      : undefined;
-
     if (queryRef.current) {
       queryRef.current.dispose();
     }
@@ -102,10 +120,12 @@ const EventsScene = ({ eventsFragmentRef }: Props) => {
     queryRef.current = refetch(
       {
         filters: {
-          ...filters,
-          interests: interestsSlug,
+          interests,
+          query,
           startDate,
           endDate,
+          minPrice,
+          maxPrice,
         },
         cursor: "0",
       },
@@ -117,19 +137,16 @@ const EventsScene = ({ eventsFragmentRef }: Props) => {
         },
       }
     );
-  }, [filters, interestsSlug]);
+  }, [query, startDate, endDate, minPrice, maxPrice, interests]);
 
   React.useEffect(() => {
     setContent(
       <Sidebar
         eventFiltersFragment={data?.eventFilters}
         interestsQueryFragmentRef={interestsQuery}
-        onChange={(args) => {
-          setFilters(args);
-        }}
       />
     );
-  }, [data, interestsQuery, filters, setFilters]);
+  }, [data, interestsQuery]);
 
   return (
     <Grid container>
@@ -138,9 +155,6 @@ const EventsScene = ({ eventsFragmentRef }: Props) => {
           <Sidebar
             eventFiltersFragment={data?.eventFilters}
             interestsQueryFragmentRef={interestsQuery}
-            onChange={(args) => {
-              setFilters(args);
-            }}
           />
         </Grid>
       )}
@@ -159,10 +173,10 @@ const EventsScene = ({ eventsFragmentRef }: Props) => {
         <Grid xl={9} lg={12} xs={12}>
           <SearchBar />
         </Grid>
-        {interestsSlug.length > 0 && (
+        {interests.length > 0 && (
           <Grid xl={9} lg={12} xs={12} p={1}>
             <Stack direction="row" useFlexGap spacing={2}>
-              {interestsSlug.map((interest: string) =>
+              {interests.map((interest: string) =>
                 interest ? (
                   <Chip
                     key={interest}
@@ -172,7 +186,7 @@ const EventsScene = ({ eventsFragmentRef }: Props) => {
                       <ChipDelete
                         onDelete={() => {
                           updateInterests(
-                            interestsSlug.filter(
+                            interests.filter(
                               (slug: string) => slug !== interest
                             )
                           );
@@ -187,18 +201,20 @@ const EventsScene = ({ eventsFragmentRef }: Props) => {
             </Stack>
           </Grid>
         )}
-        <Grid xl={9} lg={12} xs={12} container spacing={2}>
+        <Grid xl={9} lg={12} xs={12} container rowSpacing={1} gap={1} pt="20px">
           {events.map((event, index) => {
             if (index === 0) {
               if (isVeryLargeDisplay || isLargeDisplay) {
                 return (
-                  <Grid key={event.id} xs={12} lg={12} xl={12} padding={0}>
-                    <EventCardWide eventFragmentRef={event} />
+                  <Grid key={event!.id} xs={12} lg={12} xl={12}>
+                    <EventCardWide eventFragmentRef={event!} />
                   </Grid>
                 );
               }
             }
-            return <EventCardCompact key={event.id} eventFragmentRef={event} />;
+            return (
+              <EventCardCompact key={event!.id} eventFragmentRef={event!} />
+            );
           })}
           <Grid xs={12}>
             <Pagination
