@@ -6,29 +6,28 @@ import { Moment } from "moment/moment";
 import React from "react";
 import { graphql, useFragment } from "react-relay";
 import { useTranslation } from "react-i18next";
-import { Sidebar_EventFiltersFragment$key } from "../../../../../artifacts/Sidebar_EventFiltersFragment.graphql";
-import { Sidebar_InterestsFragment$key } from "../../../../../artifacts/Sidebar_InterestsFragment.graphql";
-import DateRangePicker from "../../../../../components/v2/DateRangePicker/DateRangePicker";
-import Input from "../../../../../components/v2/Input/Input";
-import Link from "../../../../../components/v2/Link";
-import SliderRange from "../../../../../components/v2/SliderRange/SliderRange";
+import moment from "moment";
+import DateRangePicker from "components/v2/DateRangePicker/DateRangePicker";
+import Input from "components/v2/Input/Input";
+import Link from "components/v2/Link";
+import SliderRange from "components/v2/SliderRange/SliderRange";
+import { parseValue, useQuery, useUpdateQuery } from "lib/hooks/useQuery";
+import { Sidebar_InterestsFragment$key } from "artifacts/Sidebar_InterestsFragment.graphql";
+import { Sidebar_EventFiltersFragment$key } from "artifacts/Sidebar_EventFiltersFragment.graphql";
+import QueryInput from "components/shared/QueryInput";
 import InterestsSelect from "./InterestsSelect";
-import { useQuery } from "../../../../../lib/hooks/useQuery";
 
 interface Props {
   eventFiltersFragment: Sidebar_EventFiltersFragment$key;
   interestsQueryFragmentRef: Sidebar_InterestsFragment$key;
-  onChange: (args: Record<string, any>) => void;
   sidebar?: boolean;
 }
 
 const Sidebar = ({
   eventFiltersFragment,
   interestsQueryFragmentRef,
-  onChange,
   sidebar,
 }: Props) => {
-  const query = useQuery("query");
   const { t } = useTranslation();
   const theme = useTheme();
   const isMediumDisplay = useMediaQuery(theme.breakpoints.up("md"));
@@ -56,42 +55,48 @@ const Sidebar = ({
     `,
     interestsQueryFragmentRef
   );
-  const ref = React.useRef<NodeJS.Timeout | null>(null);
-  const [selectedDates, setDates] = React.useState<
-    [Moment | null, Moment | null]
-  >([null, null]);
 
-  const [priceRange, setPriceRange] = React.useState<number[]>([
-    edgeFiltersValues.minPrice.cents / 100,
-    edgeFiltersValues.maxPrice.cents / 100,
-  ]);
-  const [city, setCity] = React.useState("");
-  const filters = React.useMemo(
-    () => ({
-      filters: {
-        startDate: selectedDates[0]?.toDate(),
-        endDate: selectedDates[1]?.toDate(),
-        minPrice: priceRange[0] * 100,
-        maxPrice: priceRange[1] * 100,
-        city,
-        query,
-      },
-    }),
-    [selectedDates, priceRange, city, query]
+  const priceBoundaries = {
+    min: edgeFiltersValues.minPrice.cents / 100,
+    max: edgeFiltersValues.maxPrice.cents / 100,
+  };
+
+  const minPrice = useQuery("minPrice", priceBoundaries.min, (value) =>
+    parseInt(value, 10)
   );
 
+  const maxPrice = useQuery("maxPrice", priceBoundaries.max, (value) =>
+    parseInt(value, 10)
+  );
+
+  const dates = useQuery("dates", null, (datesStr) => {
+    const datesArray = parseValue(datesStr) as string[];
+    return datesArray
+      .map((dt) => moment(dt))
+      .filter((dt: Moment) => dt.isValid());
+  });
+  const setMinPrice = useUpdateQuery("minPrice");
+  const setMaxPrice = useUpdateQuery("maxPrice");
+  const setQueryDate = useUpdateQuery("dates", (dts: Moment[]) =>
+    JSON.stringify(dts.map((val) => val.format("YYYY-MM-DD")))
+  );
+
+  const [selectedDates, setDates] =
+    React.useState<[Moment | null, Moment | null]>(dates);
+
+  const [priceRange, setPriceRange] = React.useState<number[]>([
+    minPrice,
+    maxPrice,
+  ]);
+
   React.useEffect(() => {
-    if (ref.current) {
-      clearTimeout(ref.current);
-
-      ref.current = null;
+    if (priceRange[0] !== minPrice) {
+      setMinPrice(priceRange[0]);
     }
-    ref.current = setTimeout(() => {
-      onChange(filters);
-
-      ref.current = null;
-    }, 1000);
-  }, [filters, ref]);
+    if (priceRange[1] !== maxPrice) {
+      setMaxPrice(priceRange[1]);
+    }
+  }, [priceRange, minPrice, maxPrice]);
 
   return (
     <Grid
@@ -102,26 +107,32 @@ const Sidebar = ({
       {!isMediumDisplay && (
         <Grid xs={12}>
           <Link href="/">
-            <Image src="https://placehold.co/250x75" width={300} height={90} />
+            <Image
+              src="https://placehold.co/250x75"
+              alt="logo"
+              width={300}
+              height={90}
+            />
           </Link>
         </Grid>
       )}
       <Grid xs={12}>
-        <Input
-          onChange={(value) => setCity(value)}
-          value={city}
+        <QueryInput
+          queryName="city"
           label={t("scenes.attendees.events.eventsScene.sidebar.city")}
           endDecorator={<EditIcon />}
         />
       </Grid>
       <DateRangePicker
         value={selectedDates}
-        onChange={(dates) => {
+        onChange={(dts) => {
           if (
-            dates.filter(Boolean).length === 2 ||
-            dates.filter(Boolean).length === 0
+            dts.filter(Boolean).length === 2 ||
+            dts.filter(Boolean).length === 0
           ) {
-            setDates(dates);
+            setDates(dts);
+
+            setQueryDate(dts.filter(Boolean));
           }
         }}
         minDate={edgeFiltersValues.startDate}
@@ -148,9 +159,11 @@ const Sidebar = ({
             t("scenes.attendees.events.eventsScene.sidebar.priceRance")
           }
           value={priceRange}
-          onChange={(value) => setPriceRange(value)}
-          max={edgeFiltersValues.maxPrice.cents / 100}
-          min={edgeFiltersValues.minPrice.cents / 100}
+          onChange={(value) => {
+            setPriceRange(value);
+          }}
+          max={priceBoundaries.max}
+          min={priceBoundaries.min}
           valueLabelDisplay="auto"
           size="lg"
           label={t("scenes.attendees.events.eventsScene.sidebar.priceRance")}
@@ -163,11 +176,19 @@ const Sidebar = ({
             type="number"
             value={priceRange[0].toString()}
             onChange={(value) => {
-              let newValue = parseInt(value, 10) * 100;
+              let newValue = parseInt(value, 10);
               if (Number.isNaN(newValue)) newValue = 0;
               setPriceRange([newValue, priceRange[1]]);
+
+              setMinPrice(newValue);
             }}
             label={t("scenes.attendees.events.eventsScene.sidebar.minPrice")}
+            slotProps={{
+              input: {
+                ...priceBoundaries,
+                step: 1,
+              },
+            }}
           />
         </Grid>
         <Grid xs={6}>
@@ -176,11 +197,19 @@ const Sidebar = ({
             type="number"
             value={priceRange[1].toString()}
             onChange={(value) => {
-              let newValue = parseInt(value, 10) * 100;
+              let newValue = parseInt(value, 10);
               if (Number.isNaN(newValue)) newValue = 0;
               setPriceRange([priceRange[0], newValue]);
+
+              setMaxPrice(newValue);
             }}
             label={t("scenes.attendees.events.eventsScene.sidebar.maxPrice")}
+            slotProps={{
+              input: {
+                ...priceBoundaries,
+                step: 1,
+              },
+            }}
           />
         </Grid>
       </Grid>
