@@ -2,9 +2,7 @@ import React from "react";
 import { graphql, useFragment } from "react-relay";
 import { Grid } from "@mui/joy";
 import moment, { Moment } from "moment";
-import { useTranslation } from "react-i18next";
 import Typography from "../../../../components/v2/Typography";
-import Tag from "../../../../components/v2/Tag";
 import { TripScene_TripFragment$key } from "../../../../artifacts/TripScene_TripFragment.graphql";
 import DateBookingsSection from "./components/DateBookingsSection";
 
@@ -13,7 +11,6 @@ interface TripSceneProps {
 }
 
 const TripScene = ({ tripFragmentRef }: TripSceneProps) => {
-  const { t } = useTranslation();
   const trip = useFragment(
     graphql`
       fragment TripScene_TripFragment on Trip {
@@ -50,24 +47,26 @@ const TripScene = ({ tripFragmentRef }: TripSceneProps) => {
   );
 
   return (
-    <Grid container spacing={2} padding={2}>
-      <Grid xs={12}>
-        <Typography level="h2">{trip.cities.join(", ")}</Typography>
+    <React.Suspense>
+      <Grid container spacing={2} padding={2}>
+        <Grid xs={12}>
+          <Typography level="h2">{trip.cities.join(", ")}</Typography>
+        </Grid>
+        <Grid xs={12}>
+          <Typography level="h3">
+            {moment(trip.startDate).calendar()} -{" "}
+            {moment(trip.endDate).calendar()}
+          </Typography>
+        </Grid>
+        {dates.map((dt) => (
+          <DateBookingsSection
+            key={dt.toISOString()}
+            tripFragmentRef={trip}
+            date={dt}
+          />
+        ))}
       </Grid>
-      <Grid xs={12}>
-        <Typography level="h3">
-          {moment(trip.startDate).calendar()} -{" "}
-          {moment(trip.endDate).calendar()}
-        </Typography>
-      </Grid>
-      {dates.map((dt) => (
-        <DateBookingsSection
-          key={dt.toISOString()}
-          tripFragmentRef={trip}
-          date={dt}
-        />
-      ))}
-    </Grid>
+    </React.Suspense>
   );
 };
 
