@@ -2,7 +2,6 @@ import React from "react";
 import { Moment } from "moment";
 import { graphql, useFragment } from "react-relay";
 import { Grid, Stack } from "@mui/joy";
-import { useRouter } from "next/router";
 import Typography from "../../../../../components/v2/Typography/Typography";
 import { dayMonthFormat } from "../../../../../lib/utils/dates";
 import BookingCard from "./BookingCard";
@@ -17,6 +16,11 @@ const DateBookingsSection = ({
   tripFragmentRef,
   date,
 }: DateBookingsSectionProps) => {
+  const [serverSide, setServerSide] = React.useState(true);
+  React.useEffect(() => {
+    setServerSide(false);
+  }, []);
+
   const trip = useFragment(
     graphql`
       fragment DateBookingsSection_TripFragment on Trip {
@@ -37,13 +41,20 @@ const DateBookingsSection = ({
   );
 
   React.useEffect(() => {
+    if (serverSide) {
+      return;
+    }
+
     const domNode = window.document.getElementById(
       window.document.location.hash.replace("#", "")
     );
+
     if (domNode) {
       domNode.scrollIntoView();
     }
-  }, []);
+  }, [serverSide]);
+
+  if (serverSide) return null;
 
   return (
     <Grid xs={12}>
