@@ -2,11 +2,11 @@ import React from "react";
 import { graphql, useFragment } from "react-relay";
 import moment from "moment";
 import { Grid, Stack } from "@mui/joy";
-import Typography from "../../../../components/v2/Typography";
-import TripCard from "./components/TripCard";
-import { TripsScene_AccountFragment$key } from "../../../../artifacts/TripsScene_AccountFragment.graphql";
-import Section from "../../../../components/v2/Section";
 import { useTranslation } from "react-i18next";
+import Typography from "components/v2/Typography";
+import { TripsScene_AccountFragment$key } from "artifacts/TripsScene_AccountFragment.graphql";
+import Section from "components/v2/Section";
+import TripCard from "./components/TripCard";
 
 interface TripsSceneProps {
   accountFragmentRef: TripsScene_AccountFragment$key;
@@ -27,9 +27,7 @@ const TripsScene = ({ accountFragmentRef }: TripsSceneProps) => {
     `,
     accountFragmentRef
   );
-
-  const trips = data?.trips || [];
-
+  const trips = data?.trips;
   const draftTrips = React.useMemo(
     () => trips.filter((trip) => trip.status === "draft"),
     [trips]
@@ -65,19 +63,15 @@ const TripsScene = ({ accountFragmentRef }: TripsSceneProps) => {
       ),
     [trips]
   );
-
-  const cancelledTrips = React.useMemo(
-    () => trips.filter((trip) => trip.status === "cancelled"),
-    [trips]
-  );
-
-  const { t } = useTranslation()
+  const { t } = useTranslation();
 
   return (
     <Grid>
-      <Section color={'primary'} variant='soft'>
+      <Section color="primary" variant="soft">
         <Grid xs={12}>
-          <Typography level="h3">{t('scenes.attendees.trips.tripsScene.draft')}</Typography>
+          <Typography level="h3">
+            {t("scenes.attendees.trips.tripsScene.draft")}
+          </Typography>
         </Grid>
         <Grid xs={12}>
           <Stack flexDirection="row" flexWrap="wrap">
@@ -88,9 +82,11 @@ const TripsScene = ({ accountFragmentRef }: TripsSceneProps) => {
         </Grid>
       </Section>
 
-      <Section color={'primary'} variant='soft'>
+      <Section color="primary" variant="soft">
         <Grid xs={12}>
-          <Typography level="h3">{t('scenes.attendees.trips.tripsScene.active')}</Typography>
+          <Typography level="h3">
+            {t("scenes.attendees.trips.tripsScene.active")}
+          </Typography>
         </Grid>
         <Grid xs={12}>
           <Stack>
@@ -101,9 +97,11 @@ const TripsScene = ({ accountFragmentRef }: TripsSceneProps) => {
         </Grid>
       </Section>
 
-      <Section color={'primary'} variant='soft'>
+      <Section color="primary" variant="soft">
         <Grid xs={12}>
-          <Typography level="h3">{t('scenes.attendees.trips.tripsScene.future')}</Typography>
+          <Typography level="h3">
+            {t("scenes.attendees.trips.tripsScene.future")}
+          </Typography>
         </Grid>
         <Grid xs={12}>
           <Stack>
@@ -114,9 +112,11 @@ const TripsScene = ({ accountFragmentRef }: TripsSceneProps) => {
         </Grid>
       </Section>
 
-      <Section color={'neutral'} variant='soft'>
+      <Section color="neutral" variant="soft">
         <Grid xs={12}>
-          <Typography level="h3" color='neutral'>{t('scenes.attendees.trips.tripsScene.past')}</Typography>
+          <Typography level="h3" color="neutral">
+            {t("scenes.attendees.trips.tripsScene.past")}
+          </Typography>
         </Grid>
         <Grid xs={12}>
           <Stack>
