@@ -18,23 +18,23 @@ import { graphql, useFragment } from "react-relay";
 import { useTranslation } from "react-i18next";
 import LaunchIcon from "@mui/icons-material/Launch";
 import { MoreVert } from "@mui/icons-material";
-import { EventScene_FirmEventFragment$key } from "../../../../artifacts/EventScene_FirmEventFragment.graphql";
-import GeneralInformation from "./components/GeneralInformation";
-import Typography from "../../../../components/v2/Typography";
-import EventOptionsInformation from "./components/EventOptionsInformation";
-import SchedulesInformation from "./components/SchedulesInformation";
-import Tag from "../../../../components/v2/Tag";
-import { EventScene_CurrentUserFragment$key } from "../../../../artifacts/EventScene_CurrentUserFragment.graphql";
-import VerifyEvent from "../../../../components/shared/VerifyEvent";
-import useStatusColor from "../../../../lib/hooks/useStatusColor";
-import Link from "../../../../components/v2/Link";
+import { EventScene_FirmEventFragment$key } from "artifacts/EventScene_FirmEventFragment.graphql";
+import Typography from "components/v2/Typography";
+import Tag from "components/v2/Tag";
+import { EventScene_CurrentUserFragment$key } from "artifacts/EventScene_CurrentUserFragment.graphql";
+import VerifyEvent from "components/shared/VerifyEvent";
+import useStatusColor from "lib/hooks/useStatusColor";
+import Link from "components/v2/Link";
+import Button from "components/v2/Button";
+import PublishEvent from "components/shared/PublishEvent";
+import UnpublishEvent from "components/shared/UnpublishEvent";
+import RemoveEvent from "components/shared/RemoveEvent";
+import RescheduleEvent from "components/shared/RescheduleEvent";
+import SyncStripe from "components/shared/SyncStripe";
 import BookingsInformation from "./components/BookingsInformation";
-import Button from "../../../../components/v2/Button";
-import PublishEvent from "../../../../components/shared/PublishEvent";
-import UnpublishEvent from "../../../../components/shared/UnpublishEvent";
-import RemoveEvent from "../../../../components/shared/RemoveEvent";
-import RescheduleEvent from "../../../../components/shared/RescheduleEvent";
-import SyncStripe from "../../../../components/shared/SyncStripe";
+import SchedulesInformation from "./components/SchedulesInformation";
+import EventOptionsInformation from "./components/EventOptionsInformation";
+import GeneralInformation from "./components/GeneralInformation";
 import StripeIntegrationsInformation from "./components/StripeIntegrationsInformation";
 
 interface EventSceneProps {
@@ -49,6 +49,9 @@ const EventScene = ({
   const event = useFragment<EventScene_FirmEventFragment$key>(
     graphql`
       fragment EventScene_FirmEventFragment on Event {
+        firm {
+          title
+        }
         eventOptions {
           id
         }
@@ -162,6 +165,7 @@ const EventScene = ({
             <LaunchIcon />
           </Link>
         </Typography>
+        <Typography level="h4">{event.firm.title}</Typography>
         <Tag color={tagColor} link={false}>
           {t(`statuses.${event.status}`)}
         </Tag>
