@@ -1,9 +1,9 @@
 import React from "react";
-import { Grid, Stack, TabPanel } from "@mui/joy";
+import { Grid, TabPanel } from "@mui/joy";
 import { graphql, useFragment } from "react-relay";
-import Typography from "../../../../../components/v2/Typography/Typography";
-import { BookingsInformation_EventFragment$key } from "../../../../../artifacts/BookingsInformation_EventFragment.graphql";
-import BookingsEventTable from "../../../../../components/shared/tables/BookingsEventTable";
+import Typography from "components/v2/Typography/Typography";
+import { BookingsInformation_EventFragment$key } from "artifacts/BookingsInformation_EventFragment.graphql";
+import BookingsEventTable from "components/shared/tables/BookingsEventTable";
 
 interface BookingsInformationProps {
   eventFragmentRef: BookingsInformation_EventFragment$key;
