@@ -18,11 +18,12 @@ class GraphqlController < ApplicationController
       current_user: @current_user || User.create_temporary,
       cookies: cookies
     }
+
     set_locale
     update_user_locale
     result = GraphqlSchema.execute(query, variables: variables, context: context, operation_name: operation_name)
     status_code = result&.dig('errors', 0, 'extensions', 'statusCode') || 200
-    cookies.encrypted[Stopover::AuthorizationSupport::COOKIE_KEY] = context[:current_user]&.access_token
+    cookies[Stopover::AuthorizationSupport::COOKIE_KEY] = context[:current_user]&.access_token
 
     render json: result, status_code: status_code
   rescue StandardError => e
