@@ -5,7 +5,7 @@ module Mutations
     class SignOut < BaseMutation
       field :signed_out, Boolean
       def resolve
-        context[:cookies].encrypted[Stopover::AuthorizationSupport::COOKIE_KEY] = nil
+        context[:cookies][Stopover::AuthorizationSupport::COOKIE_KEY] = nil
         context[:current_user] = nil
 
         {
