@@ -1,25 +1,52 @@
 # frozen_string_literal: true
 
 class BookingQuery
+  PER_PAGE = 30
+
   def initialize(
     params = {},
-    relations = Booking.joins(:schedule).where('scheduled_for > ?', Time.zone.now),
-    current_user = nil
+    after: 0,
+    limit: PER_PAGE,
+    backend: false
   )
-    @relations = relations
+    @backend = backend
     @params = params
-    @current_user = current_user
+    @conditions = {}
+    @offset = after
+    @limit = limit
+  end
+
+  def execute(offset: 0, limit: @limit)
+    if query && !@backend
+      Booking.search(query, where: conditions, offset: offset, limit: limit)
+    else
+      Booking.search(where: conditions, offset: offset, limit: limit)
+    end
   end
 
   def all
-    @relations = @relations.where(status: @params[:status]) if @params[:status].present?
-    @relations = @relations.where(trip_id: @params[:trip_id]) if @params[:trip_id].present?
-    if @params[:scheduled_for].present?
-      @relations = @relations.joins(:schedule).where(schedule: {
-                                                       scheduled_for: @params[:scheduled_for]
-                                                     })
-    end
-    @relations = @relations.where(event_id: @params[:event_id]) if @params[:event_id].present?
-    @relations
+    execute(offset: @offset).to_a
+  end
+
+  def total
+    execute(offset: nil, limit: nil).count
+  end
+
+  def conditions
+    @conditions[:status] = @params[:status] if @params[:status].present?
+    @conditions[:trip_id] = @params[:trip_id] if @params[:trip_id].present?
+    @conditions[:booked_for] = @params[:booked_for] if @params[:booked_for].present?
+    @conditions[:event_id] = @params[:event_id] if @params[:event_id].present?
+    @conditions[:schedule_id] = @params[:schedule_id] if @params[:schedule_id].present?
+
+    Rails.logger.debug @conditions
+
+    @conditions
+  end
+
+  def query
+    @query ||= @params[:query].nil? || @params[:query]&.empty? ? nil : @params[:query]
+
+    @query
   end
 end
