@@ -29,17 +29,17 @@ class EventsQuery
   end
 
   def total
-    execute(offset: nil, limit: nil).count
+    execute(offset: 0, limit: nil).count
   end
 
   def conditions
-    @conditions[:dates] = { gte: @params[:start_date], lte: @params[:end_date] } if @params[:start_date].present? && @params[:end_date].present?
-    @conditions[:price] = { gte: @params[:min_price], lte: @params[:max_price] } if @params[:min_price].present? && @params[:max_price].present?
-    @conditions[:city] = @params[:city] if @params[:city].present? && !@params[:city].empty?
+    @conditions[:dates] =     { gte: @params[:start_date], lte: @params[:end_date] } if @params[:start_date].present? && @params[:end_date].present?
+    @conditions[:price] =     { gte: @params[:min_price] * 100, lte: @params[:max_price] * 100 } if @params[:min_price].present? && @params[:max_price].present?
+    @conditions[:city] =      @params[:city] if @params[:city].present? && !@params[:city].empty?
     @conditions[:interests] = @params[:interests] if @params[:interests]&.any?
-    @conditions[:firm_id] = @params[:firm].id if @params[:firm].present?
+    @conditions[:firm_id] =   @params[:firm_id] if @params[:firm_id].present?
 
-    @conditions[:title] = query if @backend && query
+    @conditions[:title] =     query if @backend && query
 
     @conditions
   end
