@@ -28,7 +28,8 @@ class SchedulesQuery
 
   def conditions
     @conditions[:scheduled_for] = { gte: @params[:start_date], lte: @params[:end_date] } if @params[:start_date].present? && @params[:end_date].present?
-    @conditions[:event_id] = @params[:event_ids] if @params[:event_ids].present?
+    @conditions[:firm_id] = @params[:firm_id] if @params[:firm_id].present?
+    @conditions[:event_ids] = @params[:event_ids] if @params[:event_ids].present?
 
     @conditions
   end
